Refactor testing structure and consolidate test files. Removed obsolete test files for candidate validation, service, and controller, and updated Jest configuration for both frontend and backend. Introduced a new comprehensive test file for candidate data validation and service logic, ensuring all necessary tests are included in a single location. Updated package.json scripts for improved test execution.
