--- frontend/src/__tests__/App.test.tsx
@@ -1,9 +1,0 @@
-import { render, screen } from '@testing-library/react';
-import App from '../App';
-
-describe('App', () => {
-  it('renders without crashing', () => {
-    render(<App />);
-    expect(screen).toBeDefined();
-  });
-});
